pipeline_test/requirements.txt
fastapi
uvicorn
tensorflow
pillow
numpy
requests
boto3
python-dotenv
openai
scipy